refactor(runtime-common): move realm index card helpers into dedicated module
